completed authorization

+++ node-backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post("/", checkauth, upload.single("image"), (req, res, next) => {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -92,23 +93,24 @@ router.put("/:id", checkauth, upload.single("image"), (req, res, next) => {
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
+    creator: req.userData.userId,
     imagePath
   })
-  console.log("post...", post)
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    // if (result.modifiedCount)
-    res.status(200).json({ message: "Update successful !" })
-    // else
-    //   res.status(500).json({ message: "Problem in Updating !!" })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.modifiedCount)
+      res.status(200).json({ message: "Update successful !" })
+    else
+      res.status(500).json({ message: "Not Authorized !!" })
   })
 })
 
 router.delete("/:id", checkauth, async (req, res) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
     if (result.deletedCount)
-      res.status(200).json({
-        message: "Post Deleted!!"
-      })
+      res.status(200).json({ message: "Post Deleted!!" })
+    else
+      res.status(500).json({ message: "Not Authorized !!" })
+
   })
 
 })
